refactor: LikeButton component; change icons path
