+++ src/data/configMenu/configMenu.tsx
@@ -1,11 +1,11 @@
-import HomeSVG from '@/assets/Home.svg';
-import ExplorerSVG from '@/assets/Explore.svg';
-import NotificationSVG from '@/assets/Notifications.svg';
-import MessagesSVG from '@/assets/Messages.svg';
-import BookmarksSVG from '@/assets/Bookmarks.svg';
-import ListsSVG from '@/assets/Lists.svg';
-import ProfileSVG from '@/assets/Profile.svg';
-import MoreSVG from '@/assets/More.svg';
+import HomeSVG from '@icons/Home.svg';
+import ExplorerSVG from '@/assets/icons/Explore.svg';
+import NotificationSVG from '@/assets/icons/Notifications.svg';
+import MessagesSVG from '@/assets/icons/Messages.svg';
+import BookmarksSVG from '@/assets/icons/Bookmarks.svg';
+import ListsSVG from '@/assets/icons/Lists.svg';
+import ProfileSVG from '@/assets/icons/Profile.svg';
+import MoreSVG from '@/assets/icons/More.svg';
 
 const currentPageColor = '#1DA1F2';
 const anotherPagesColor = '#0F1419';
